Improve condition for determining best solution
Previously using end value and stopping on finding a root, this favours
diverging solutions (which are harder to optimise on), as opposed to
those whose sign flips (the max value of v_θ is the ideal optimisation
parameter for this). Now use max value and no root stopping, so we mark
down solutions which jump around like crazy.